+++ app/src/main/java/com/example/newapppp/data/remote/modul/GetHabitJson.kt
@@ -15,6 +15,6 @@ data class GetHabitJson(
     val title: String,
     val type: Int,
     @SerialName(value = "done_dates")
-    val doneDates: List<Int>,
+    val doneDates: List<Long>,
     val uid: String
 )
