+++ app/src/main/java/com/example/newapppp/domain/model/HabitSave.kt
@@ -11,6 +11,6 @@ data class HabitSave(
     val color : HabitColor,
     val priority : HabitPriority,
     val type : HabitType,
-    val doneDates: List<Int>,
+    val doneDates: List<Long>,
     val frequency : Int
 )
